+++ app/src/main/java/ru/practicum/android/diploma/features/filter/setting/composables/SalaryTextField.kt
@@ -1,12 +1,7 @@
 package ru.practicum.android.diploma.features.filter.setting.composables
 
-import android.app.Activity
-import android.content.Context
-import android.view.inputmethod.InputMethodManager
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.focusable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -17,7 +12,6 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -34,7 +28,6 @@ import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.graphics.SolidColor
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
@@ -57,7 +50,6 @@ fun SalaryTextField(
 ) {
     val focusRequester = remember { FocusRequester() }
     var isFocused by remember { mutableStateOf(false) }
-    val context = LocalContext.current
     val focusManager = LocalFocusManager.current
 
     val textColor = if (isFocused) {
@@ -135,28 +127,7 @@ fun SalaryTextField(
                     }
                 }
             }
-
-            if (query.isNotEmpty()) {
-                Icon(
-                    modifier = Modifier
-
-                        .clickable(
-                            interactionSource = MutableInteractionSource(),
-                            indication = null,
-                        ) {
-                            val imm = context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
-                            val view = (context as? Activity)?.currentFocus
-                            view?.let {
-                                imm.hideSoftInputFromWindow(it.windowToken, 0)
-                            }
-                            onValueChange("")
-                            viewModel.changeSalary(null)
-                        },
-                    painter = painterResource(id = R.drawable.close),
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.onSurface
-                )
-            }
+            ClearIcon(query, onValueChange, viewModel)
         }
     }
 }
